nutrition-label-agent/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain & LangGraph
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40
langchain-community>=0.0.20

# Tavily Search
tavily-python>=0.3.0

# Image processing
Pillow>=10.0.0

# Pydantic
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0
